+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * FIXED:
     * Core: do not load the WPGlobus translations from WordPress.org and always use those from the plugin's `languages` folder.
 	* Customizer: fixed warning "strpos() expects parameter 1 to be string, array given".
+* ADDED:
+    * Recommend `WPGlobus Plus` to edit permalinks.
 
 ### 1.9.5 ###